requirements.txt
future==0.15.2

Django==1.9.7
click==6.2
django-redis-cache==1.6.5
-e git+https://github.com/ui/django-rq.git@3bd07acd92143fa4e77089b6d65cde5a7a6aee4b#egg=django_rq-master
redis==2.10.5
rq==0.5.6
pytz==2015.7
python-dateutil==2.5.3

six==1.10.0
django-environ==0.4.0

requests==2.9.1

py==1.4.31
pytest==2.8.7
pytest-django==2.9.1
flake8==2.6.2
mccabe==0.5.0
pycodestyle==2.0.0
pyflakes==1.2.3
pytest-flake8==0.5


coverage==4.1
pytest-cov==2.3.0
pytest-cover==3.0.0
pytest-coverage==0.0
